// Front_End/src/theme.d.ts
declare module '@/src/theme' {
  export const colors: {
    logo: string;
    black: string;
    white: string;
  };

  export const typography: {
    fontFamily: {
      madimi: string;
      semiBold: string;
      medium: string;
    };
    fontSize: {
      XLarge: number;
      XXLarge: number;
      body: number;
    };
    lineHeight: {
      body: number;
    };
    letterSpacing: {
      normal: number;
    };
  };
} 